added create meeting functionality

+++ simpleAttendance2.js
@@ -33,11 +33,58 @@ $(function () {
     //delete an employee
     $('#employeeList').on('click', '.removeEmployee', function (e) {
        e.preventDefault();
+
+       let delEmpResponse = confirm('Are you sure you want to delete employee?');
+       if (delEmpResponse) {
+        $.ajax({
+            type: 'DELETE',
+            url: 'http://localhost:3000/employeeList/' + $(this).attr("data-id"), 
+        });
+       };
+    });
+
+    //create a meeting
+    $('#date').datepicker();
+
+    $('#createMeeting').click(function (e) {
+       e.preventDefault();
        
-       $.ajax({
-        type: 'DELETE',
-        url: 'http://localhost:3000/employeeList/' + $(this).attr("data-id"), 
-       });
-       
+       const meetingDate = $('#date').val();
+       const meetingSubject = $('#subject').val(); 
+
+       const newMeeting = {
+        date: meetingDate,
+        subject: meetingSubject
+       }
+
+       if (meetingDate === '' || meetingSubject === '' ) {
+        alert('Please fill both fields')
+        } else {
+            $.post('http://localhost:3000/meeting', newMeeting, alert('Meeting has been added'))
+        }
+    });
+
+    //get employee  names to be added to meeting
+    $.get("http://localhost:3000/employeeList", function(employeeInfo){
+        $.each(employeeInfo, function(i){
+            $('#meetingEmployee').append(`<li><input type="checkbox" value="${employeeInfo[i].name}">${employeeInfo[i].name} 
+            <div>
+                <p>Attendance</p>
+                <div>
+                    <input type="radio" name="attendance${employeeInfo[i].id}" id="attendancePresent" value="Present">
+                    <label>Present</label>
+                    <input type="radio" name="attendance${employeeInfo[i].id}" id="attendanceAbsent" value="Absent">
+                    <label>Absent</label>
+                </div>
+            </div>
+            </li>`)
+        });
+    });
+    
+    //get meeting from server 
+    $.get("http://localhost:3000/meeting", function(selectMeeting){
+        $.each(selectMeeting, function(i){
+            $('#selectMeeting').append(`<option value="${selectMeeting[i].subject}">${selectMeeting[i].subject}</option>`)
+        });
     });
 });
